SpringConfig 제거, id 타입을 int 로 변경
+ 굳이 @Bean을 사용할 필요성을 못 느껴서 @Service, @Repository를 사용하는걸로 변경
+ id 값 또한 지금으로선 Long은 오버킬 같아서 int로 변경

+++ src/main/java/com/kakao/cafe/controller/ArticleController.java
@@ -32,7 +32,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{index}")
-    public String view(@PathVariable Long index, Model model) {
+    public String view(@PathVariable int index, Model model) {
         Article article = articleService.findById(index);
         model.addAttribute("article", article);
